created ActiveStateBase
